Reskilling MVP_1/ai_insight_generator/backend/requirements.txt
Flask
Flask-CORS
graphene
openai
python-dotenv

